Split Fal response helpers
